업로드된 파일에 해당하는 gen_file row 생성되도록

+++ src/main/java/com/example/file/app/fileUpload/entity/GenFile.java
@@ -17,7 +17,7 @@ public class GenFile extends BaseEntity {
 
     private String relTypeCode;
 
-    private int relId;
+    private long relId;
 
     private String typeCode;
